pyproject.toml
# Poetry pyproject.toml: https://python-poetry.org/docs/pyproject/
[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "snyk-to-azure-boards"
version = "0.1.0"
description = "CLI tool to convert Snyk issues to Azure Boards issues."
readme = "README.md"
authors = ["Omar Quimbaya <omar.quimbaya@snyk.io>"]
license = "Apache Software License 2.0"
repository = "https://github.com/writingpanda/snyk-to-azure-boards"
homepage = "https://github.com/writingpanda/snyk-to-azure-boards"

# Keywords description https://python-poetry.org/docs/pyproject/#keywords
keywords = ["azure boards", "snyk"]

# Pypi classifiers: https://pypi.org/classifiers/
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
]

[tool.poetry.scripts]
# Entry points for the package https://python-poetry.org/docs/pyproject/#scripts
"snyk-to-azure-boards" = "snyk_to_azure_boards.__main__:app"

[tool.poetry.dependencies]
python = "^3.11"

typer = {extras = ["all"], version = "^0.7.0"}
rich = "^12.6.0"
pysnyk = "^0.9.4"
azure-devops = "^6.0.0b4"

[tool.poetry.dev-dependencies]
bandit = "^1.7.1"
black = {version = "^23.1a1", allow-prereleases = true}
darglint = "^1.8.1"
isort = {extras = ["colors"], version = "^5.11.4"}
mypy = "^0.991"
mypy-extensions = "^0.4.3"
pre-commit = "^2.15.0"
pydocstyle = "^6.1.1"
pylint = "^2.11.1"
pytest = "^7.2.0"
pyupgrade = "^3.3.1"
safety = "^2.3.5"
coverage = "^7.0.0"
coverage-badge = "^1.1.0"
pytest-html = "^3.1.1"
pytest-cov = "^4.0.0"

[tool.poetry.group.dev.dependencies]
bandit = "^1.7.4"
darglint = "^1.8.1"
isort = {extras = ["colors"], version = "^5.11.4"}
mypy = "^0.991"
pre-commit = "^2.20.0"
pydocstyle = "^6.1.1"
pylint = "^2.15.9"
pytest = "^7.2.0"
pyupgrade = "^3.3.1"
safety = "^2.3.5"
coverage = "^7.0.0"
coverage-badge = "^1.1.0"
pytest-html = "^3.2.0"
pytest-cov = "^4.0.0"
black = {version = "^23.1a1", allow-prereleases = true}

[tool.black]
# https://github.com/psf/black
target-version = ["py311"]
line-length = 88
color = true

exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | env
    | venv
)/
'''

[tool.isort]
# https://github.com/timothycrosley/isort/
py_version = 311
line_length = 88

known_typing = ["typing", "types", "typing_extensions", "mypy", "mypy_extensions"]
sections = ["FUTURE", "TYPING", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
include_trailing_comma = true
profile = "black"
multi_line_output = 3
indent = 4
color_output = true

[tool.mypy]
# https://mypy.readthedocs.io/en/latest/config_file.html#using-a-pyproject-toml-file
python_version = 3.11
pretty = true
show_traceback = true
color_output = true

allow_redefinition = false
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
ignore_missing_imports = true
implicit_reexport = false
no_implicit_optional = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
strict_equality = true
strict_optional = true
warn_no_return = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true


[tool.pytest.ini_options]
# https://docs.pytest.org/en/6.2.x/customize.html#pyproject-toml
# Directories that are not visited by pytest collector:
norecursedirs =["hooks", "*.egg", ".eggs", "dist", "build", "docs", ".tox", ".git", "__pycache__"]
doctest_optionflags = ["NUMBER", "NORMALIZE_WHITESPACE", "IGNORE_EXCEPTION_DETAIL"]

# Extra options:
addopts = [
  "--strict-markers",
  "--tb=short",
  "--doctest-modules",
  "--doctest-continue-on-failure",
]

[tool.coverage.run]
source = ["tests"]

[coverage.paths]
source = "snyk-to-azure-boards"

[coverage.run]
branch = true

[coverage.report]
fail_under = 50
show_missing = true
